+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # v0.50.0 (unreleased)
 - Update to rubocop v0.50.0 and rubocop-rspec v1.18.0.
+- Do not apply `Ezcater/StyleDig` to assignments with nested access.
 
 # v0.49.1
 - Add `Ezcater/RspecRequireBrowserMock` cop.